pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-brief"
version = "0.1.0"
description = "Personal intelligence briefing system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "feedparser>=6.0",
    "requests>=2.28",
    "anthropic>=0.18",
    "flask>=3.0",
    "markdown>=3.5",
    "pymdown-extensions>=10.0",
    "python-dateutil>=2.8",
]

[project.scripts]
daily-brief = "cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
